requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
